Changed to singular names to reflect update in yarn PR
