+++ fabric-data-generation-api-v1/src/main/resources/fabric-data-generation-api-v1.mixins.json
@@ -3,7 +3,7 @@
   "package": "net.fabricmc.fabric.mixin.datagen",
   "compatibilityLevel": "JAVA_16",
   "mixins": [
-    "ModelsProviderMixin",
+    "ModelProviderMixin",
     "TagBuilderMixin"
   ],
   "client": [
