+++ fabric-data-generation-api-v1/src/testmod/java/net/fabricmc/fabric/test/datagen/DataGeneratorTestEntrypoint.java
@@ -41,28 +41,28 @@ import net.minecraft.util.Identifier;
 
 import net.fabricmc.fabric.api.datagen.v1.DataGeneratorEntrypoint;
 import net.fabricmc.fabric.api.datagen.v1.FabricDataGenerator;
-import net.fabricmc.fabric.api.datagen.v1.provider.FabricAdvancementsProvider;
-import net.fabricmc.fabric.api.datagen.v1.provider.FabricBlockLootTablesProvider;
-import net.fabricmc.fabric.api.datagen.v1.provider.FabricModelsProvider;
-import net.fabricmc.fabric.api.datagen.v1.provider.FabricRecipesProvider;
-import net.fabricmc.fabric.api.datagen.v1.provider.FabricTagsProvider;
-import net.fabricmc.fabric.api.datagen.v1.provider.SimpleFabricLootTablesProvider;
+import net.fabricmc.fabric.api.datagen.v1.provider.FabricAdvancementProvider;
+import net.fabricmc.fabric.api.datagen.v1.provider.FabricBlockLootTableProvider;
+import net.fabricmc.fabric.api.datagen.v1.provider.FabricModelProvider;
+import net.fabricmc.fabric.api.datagen.v1.provider.FabricRecipeProvider;
+import net.fabricmc.fabric.api.datagen.v1.provider.FabricTagProvider;
+import net.fabricmc.fabric.api.datagen.v1.provider.SimpleFabricLootTableProvider;
 
 public class DataGeneratorTestEntrypoint implements DataGeneratorEntrypoint {
 	@Override
 	public void onInitializeDataGenerator(FabricDataGenerator dataGenerator) {
-		dataGenerator.addProvider(TestRecipesProvider::new);
-		dataGenerator.addProvider(TestModelsProvider::new);
-		dataGenerator.addProvider(TestAdvancementsProvider::new);
-		dataGenerator.addProvider(TestBlockLootTablesProvider::new);
-		dataGenerator.addProvider(TestBarterLootTablesProvider::new);
+		dataGenerator.addProvider(TestRecipeProvider::new);
+		dataGenerator.addProvider(TestModelProvider::new);
+		dataGenerator.addProvider(TestAdvancementProvider::new);
+		dataGenerator.addProvider(TestBlockLootTableProvider::new);
+		dataGenerator.addProvider(TestBarterLootTableProvider::new);
 
-		TestBlockTagsProvider blockTagsProvider = dataGenerator.addProvider(TestBlockTagsProvider::new);
-		dataGenerator.addProvider(new TestItemTagsProvider(dataGenerator, blockTagsProvider));
+		TestBlockTagProvider blockTagsProvider = dataGenerator.addProvider(TestBlockTagProvider::new);
+		dataGenerator.addProvider(new TestItemTagProvider(dataGenerator, blockTagsProvider));
 	}
 
-	private static class TestRecipesProvider extends FabricRecipesProvider {
-		private TestRecipesProvider(FabricDataGenerator dataGenerator) {
+	private static class TestRecipeProvider extends FabricRecipeProvider {
+		private TestRecipeProvider(FabricDataGenerator dataGenerator) {
 			super(dataGenerator);
 		}
 
@@ -72,8 +72,8 @@ public class DataGeneratorTestEntrypoint implements DataGeneratorEntrypoint {
 		}
 	}
 
-	private static class TestModelsProvider extends FabricModelsProvider {
-		private TestModelsProvider(FabricDataGenerator generator) {
+	private static class TestModelProvider extends FabricModelProvider {
+		private TestModelProvider(FabricDataGenerator generator) {
 			super(generator);
 		}
 
@@ -88,8 +88,8 @@ public class DataGeneratorTestEntrypoint implements DataGeneratorEntrypoint {
 		}
 	}
 
-	private static class TestBlockTagsProvider extends FabricTagsProvider.BlockTagsProvider {
-		private TestBlockTagsProvider(FabricDataGenerator dataGenerator) {
+	private static class TestBlockTagProvider extends FabricTagProvider.BlockTagProvider {
+		private TestBlockTagProvider(FabricDataGenerator dataGenerator) {
 			super(dataGenerator);
 		}
 
@@ -100,8 +100,8 @@ public class DataGeneratorTestEntrypoint implements DataGeneratorEntrypoint {
 		}
 	}
 
-	private static class TestItemTagsProvider extends FabricTagsProvider.ItemTagsProvider {
-		private TestItemTagsProvider(FabricDataGenerator dataGenerator, BlockTagsProvider blockTagProvider) {
+	private static class TestItemTagProvider extends FabricTagProvider.ItemTagProvider {
+		private TestItemTagProvider(FabricDataGenerator dataGenerator, BlockTagProvider blockTagProvider) {
 			super(dataGenerator, blockTagProvider);
 		}
 
@@ -111,8 +111,8 @@ public class DataGeneratorTestEntrypoint implements DataGeneratorEntrypoint {
 		}
 	}
 
-	private static class TestAdvancementsProvider extends FabricAdvancementsProvider {
-		private TestAdvancementsProvider(FabricDataGenerator dataGenerator) {
+	private static class TestAdvancementProvider extends FabricAdvancementProvider {
+		private TestAdvancementProvider(FabricDataGenerator dataGenerator) {
 			super(dataGenerator);
 		}
 
@@ -131,8 +131,8 @@ public class DataGeneratorTestEntrypoint implements DataGeneratorEntrypoint {
 		}
 	}
 
-	private static class TestBlockLootTablesProvider extends FabricBlockLootTablesProvider {
-		private TestBlockLootTablesProvider(FabricDataGenerator dataGenerator) {
+	private static class TestBlockLootTableProvider extends FabricBlockLootTableProvider {
+		private TestBlockLootTableProvider(FabricDataGenerator dataGenerator) {
 			super(dataGenerator);
 		}
 
@@ -142,8 +142,8 @@ public class DataGeneratorTestEntrypoint implements DataGeneratorEntrypoint {
 		}
 	}
 
-	private static class TestBarterLootTablesProvider extends SimpleFabricLootTablesProvider {
-		private TestBarterLootTablesProvider(FabricDataGenerator dataGenerator) {
+	private static class TestBarterLootTableProvider extends SimpleFabricLootTableProvider {
+		private TestBarterLootTableProvider(FabricDataGenerator dataGenerator) {
 			super(dataGenerator, LootContextTypes.BARTER);
 		}
